fix compile error
svn path=/trunk/kdenetwork/krdc/; revision=159464

+++ threads.cpp
@@ -311,7 +311,7 @@ void WriterThread::run() {
 				}
 			if (!clientCut.isNull()) 
 				if (!SendClientCutText(clientCut.latin1(), 
-						       clientCut.length())) {
+						       (int)clientCut.length())) {
 					sendFatalError(ERROR_IO);
 					break;
 				}
